refactor: unify storage layer into projectStorage facade
Merge project-manager.ts (localStorage) and asset-store.ts (IndexedDB)
into a single manager-style singleton, src/lib/project-storage.ts,
matching the existing audioManager/historyStore house style.

Pure refactoring: implementations moved verbatim, no behavior change.
The facade header documents the storage map and previously implicit
invariants (assets are not garbage-collected on project delete;
save() also bumps updatedAt and lastProjectId).

Step 1 of the app-initialization refactor
(docs/2026-01-12-app-initialization-architecture.md).
